Found a better way to handle the messages.
The messenger can setup its dispatcher to call functions directly on the
cluckd object. This is now done that way and we avoid heavy duplication
of the functions. The only drawback is that the msg_...() functions have
to be public no matter what. Since this is not a library, I'm not too
worried.