Fixed many tests and started work on the daemon.
The cluck library now sends extra COMMANDS messages so I had to update
the tests to account for that.

This also includes the new cluck daemon test, which is not 100% complete
but it is getting there. This is testing quite a few things already, though.

Added messages that we are dealing with such as the LOCK and LOCK_STARTED.