Changed directory order to test original.
By default we want to test the cluck original message definitions, not an old
version that got installed earlier. We still need the second path since we
want to test against all the message declarations from other projects.

Also added the cmake code to load and use the VerifyMessageDefinitions() macro.