Started work on test with 2 cluckd running.
This includes a test that simulates one cluckd so it looks like we have two
cluckd servers running and verify that this works as expected.

Replaced the SIGUSR1 and SIGUSR2 with an INFO message which means we can
avoid all sorts of race conditions. I've done that after fixing some of
the issues with the kill() instruction implementation in the reporter.