Found it, same as the 3 computers.
The exit() is "too violent" and require a proper shutdown of the server
first. We may have to look into fixing that issue in the reporter at
some pointer. For now, fixing the script a tad bit is sufficient and
works as expected. (i.e. wait(drain) must have a clear_message() first
to make sure we do not return immediately with the last message, this
is because we do not call the regular wait_message() which does that)