Implemented the 5-locks test.
Fixed the msg_unlock(). If an unlock happens, we must call the
activate_first_lock() function to make sure that if another lock is waiting
it gets processed and a LOCKED message gets sent.

Added a few more f_tag to messages where it was still missing.

Added a test so we detect a flood of lock requests that do not give the
cluck daemon(s) to process those requests (of form of DoS attack). This
test will reply immediately with a failure giving the daemon a chance to
process the existing lock requests.

Enhanced several errors to get more information such as the object_name
and key or tag of the concerned lock.

Enhanced documentation of many functions.

Implemented a test with 5 locks, 3 for the same object and two others.
The test passes with the current version (after a few bug fixes).

Cleaned up the test of 3 computers (i.e. removed the commented out code).

Added TODO about replies that will never make it. In all likelihood,
these will generate time outs even if the missing leaders reappear fast
enough to handle those locks. This can cause many more time outs than
just those few where a reply was missing. Although maybe the timeout
would not even work... That would be TBD.