Properly send a LOCK_READY in case we have a single computer.
When there is only one computer, we lack the exchange of the LOCK_STARTED
messages so we would not trigger the proper LOCK_READY message once the
CLUSTER_UP message was recieved.

Avoid a potential crash in some tests where we used dynamic cast and did
not verify whether the pointer was null or not.

Added RAII to reset the log counters of advgetopt (otherwise all the
calls after the error check fail).

Added a test to verify that the GET_MAX_TICKET detects overflows.

Properly handle the LOCK_READY in the one computer test.

Added the unlock_duration to a few verification tests since it is expected
to be there.