Enhanced tests further.
Fixed the sorting of computers keeping the priority as it is and only placing
leaders first by using 'A' for leaders and 'B' for other computers. That
way the priority is still taken in account as expected.

Fixed the timestamp parameter value of the ALIVE message. It's now a timespec.

Further testing losing and getting back computers.

Tested the ALIVE / ABSOLUTELY messages, although I still need to find a way
to get the sequence started on its own.