Updated daemon to send LOCK_TICKETS as expected.
The previous version missed a few calls to make sure that the LOCK_TICKETS
would be sent to the new server that becomes a leader. However, I'm not
really 100% sure that it is complete just yet. i.e. I think that the new
computer may not be ready when we try to send the LOCK_TICKETS and yet it
will be required and won't happen. So I think there is more to do.