Updated the daemon to wait for a stable clock.
This version sends a CLOCK_STATE to get one of the CLOCK_STABLE or
CLOCK_UNSTABLE. Only start the lock system if the CLOCK_STABLE
is sent and the clock was indeed properly verified.

In all other cases, just sit there until our clock becomes stable.
In most likelihood, this is an extremely fast process.

Updated the tests accordingly.